requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
pymongo==3.12.1
python-dotenv==0.19.2
Flask==2.0.3
Werkzeug==2.0.3

